[ISSUE #8236]♻️Establish the Proxy Core protocol owner
