+++ rocketmq-proxy-core/Cargo.toml
@@ -1,0 +1,50 @@
+# Copyright 2023 The RocketMQ Rust Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+[package]
+name = "rocketmq-proxy-core"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords = ["rocketmq", "proxy", "grpc", "protocol"]
+categories = ["network-programming", "asynchronous"]
+description = "Runtime-neutral RocketMQ Proxy contracts and ingress state"
+rust-version.workspace = true
+
+[features]
+default = []
+
+[dependencies]
+rocketmq-error = { workspace = true }
+rocketmq-model = { workspace = true }
+rocketmq-protocol = { workspace = true }
+rocketmq-transport = { workspace = true, default-features = false }
+
+dashmap.workspace = true
+serde.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+uuid.workspace = true
+
+prost = "0.14"
+prost-types = "0.14"
+tonic = { version = "0.14.6", features = ["transport", "codegen"] }
+tonic-prost = "0.14"
+
+[build-dependencies]
+tonic-prost-build = "0.14.6"
