+++ rocketmq-proxy-core/build.rs
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    println!("cargo:rerun-if-changed=proto/");
+    println!("cargo:rerun-if-changed=proto/definition.proto");
+    println!("cargo:rerun-if-changed=proto/service.proto");
 
     tonic_prost_build::configure().compile_protos(&["proto/service.proto"], &["proto"])?;
