+++ rocketmq-proxy-core/src/config.rs
@@ -1,0 +1,193 @@
+// Copyright 2023 The RocketMQ Rust Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::net::SocketAddr;
+use std::time::Duration;
+
+use rocketmq_error::RocketMQError;
+use serde::Deserialize;
+use serde::Serialize;
+
+use crate::ProxyResult;
+use crate::DEFAULT_PROXY_GRPC_PORT;
+use crate::DEFAULT_PROXY_REMOTING_PORT;
+
+/// Backend mode selected by the Proxy composition layer.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Default)]
+#[serde(rename_all = "snake_case")]
+pub enum ProxyMode {
+    #[default]
+    Cluster,
+    Local,
+}
+
+/// Normalized gRPC ingress configuration.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(default, rename_all = "camelCase")]
+pub struct GrpcConfig {
+    pub listen_addr: String,
+    pub max_decoding_message_size: usize,
+    pub max_encoding_message_size: usize,
+    pub concurrency_limit_per_connection: usize,
+    pub use_endpoint_port_from_request: bool,
+}
+
+impl Default for GrpcConfig {
+    fn default() -> Self {
+        Self {
+            listen_addr: format!("0.0.0.0:{DEFAULT_PROXY_GRPC_PORT}"),
+            max_decoding_message_size: 8 * 1024 * 1024,
+            max_encoding_message_size: 8 * 1024 * 1024,
+            concurrency_limit_per_connection: 256,
+            use_endpoint_port_from_request: false,
+        }
+    }
+}
+
+impl GrpcConfig {
+    pub fn socket_addr(&self) -> ProxyResult<SocketAddr> {
+        self.listen_addr.parse().map_err(|error| {
+            RocketMQError::illegal_argument(format!(
+                "invalid proxy gRPC listen address '{}': {error}",
+                self.listen_addr
+            ))
+            .into()
+        })
+    }
+
+    pub fn listen_port(&self) -> ProxyResult<u16> {
+        Ok(self.socket_addr()?.port())
+    }
+}
+
+/// Normalized RocketMQ remoting ingress configuration.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(default, rename_all = "camelCase")]
+pub struct RemotingConfig {
+    pub enabled: bool,
+    pub listen_addr: String,
+}
+
+impl Default for RemotingConfig {
+    fn default() -> Self {
+        Self {
+            enabled: false,
+            listen_addr: format!("0.0.0.0:{DEFAULT_PROXY_REMOTING_PORT}"),
+        }
+    }
+}
+
+impl RemotingConfig {
+    pub fn socket_addr(&self) -> ProxyResult<SocketAddr> {
+        self.listen_addr.parse().map_err(|error| {
+            RocketMQError::illegal_argument(format!(
+                "invalid proxy remoting listen address '{}': {error}",
+                self.listen_addr
+            ))
+            .into()
+        })
+    }
+
+    pub fn listen_port(&self) -> ProxyResult<u16> {
+        Ok(self.socket_addr()?.port())
+    }
+}
+
+/// Runtime admission limits consumed by neutral Proxy services.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(default, rename_all = "camelCase")]
+pub struct RuntimeConfig {
+    pub route_permits: usize,
+    pub producer_permits: usize,
+    pub consumer_permits: usize,
+    pub client_manager_permits: usize,
+}
+
+impl Default for RuntimeConfig {
+    fn default() -> Self {
+        Self {
+            route_permits: 512,
+            producer_permits: 1024,
+            consumer_permits: 1024,
+            client_manager_permits: 512,
+        }
+    }
+}
+
+/// Session lifecycle policy consumed by the Core registry.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(default, rename_all = "camelCase")]
+pub struct SessionConfig {
+    pub client_ttl_ms: u64,
+    pub receipt_handle_ttl_ms: u64,
+    pub auto_renew_enabled: bool,
+    pub min_long_polling_timeout_ms: u64,
+    pub max_long_polling_timeout_ms: u64,
+}
+
+impl Default for SessionConfig {
+    fn default() -> Self {
+        Self {
+            client_ttl_ms: 60_000,
+            receipt_handle_ttl_ms: 5 * 60_000,
+            auto_renew_enabled: true,
+            min_long_polling_timeout_ms: 5_000,
+            max_long_polling_timeout_ms: 20_000,
+        }
+    }
+}
+
+impl SessionConfig {
+    pub fn client_ttl(&self) -> Duration {
+        Duration::from_millis(self.client_ttl_ms.max(1))
+    }
+
+    pub fn receipt_handle_ttl(&self) -> Duration {
+        Duration::from_millis(self.receipt_handle_ttl_ms.max(1))
+    }
+
+    pub fn min_long_polling_timeout(&self) -> Duration {
+        Duration::from_millis(self.min_long_polling_timeout_ms)
+    }
+
+    pub fn max_long_polling_timeout(&self) -> Duration {
+        Duration::from_millis(self.max_long_polling_timeout_ms.max(self.min_long_polling_timeout_ms))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn ingress_defaults_preserve_public_ports() {
+        assert_eq!(GrpcConfig::default().listen_port().expect("gRPC port"), 8081);
+        assert_eq!(RemotingConfig::default().listen_port().expect("remoting port"), 8080);
+    }
+
+    #[test]
+    fn session_durations_preserve_bounds() {
+        let config = SessionConfig {
+            client_ttl_ms: 0,
+            receipt_handle_ttl_ms: 0,
+            min_long_polling_timeout_ms: 10,
+            max_long_polling_timeout_ms: 5,
+            ..SessionConfig::default()
+        };
+
+        assert_eq!(config.client_ttl(), Duration::from_millis(1));
+        assert_eq!(config.receipt_handle_ttl(), Duration::from_millis(1));
+        assert_eq!(config.max_long_polling_timeout(), Duration::from_millis(10));
+    }
+}
