+++ rocketmq-proxy/Cargo.toml
@@ -43,6 +43,7 @@ rocketmq-broker      = { workspace = true }
 rocketmq-store       = { workspace = true }
 rocketmq-error       = { workspace = true }
 rocketmq-model       = { workspace = true }
+rocketmq-proxy-core  = { workspace = true }
 rocketmq-remoting    = { workspace = true }
 rocketmq-client-rust = { workspace = true }
 rocketmq-auth        = { workspace = true, features = ["grpc"] }
@@ -66,13 +67,8 @@ cheetah-string = { workspace = true }
 thiserror      = { workspace = true }
 
 async-trait = "0.1"
-prost       = "0.14"
 prost-types = "0.14"
 tonic       = { version = "0.14.6", features = ["transport", "codegen"] }
-tonic-prost = "0.14"
-
-[build-dependencies]
-tonic-prost-build = "0.14.6"
 
 [dev-dependencies]
 hex = "0.4"
