+++ rocketmq-proxy/src/config.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use std::fmt;
-use std::net::SocketAddr;
 use std::path::Path;
 use std::time::Duration;
 
@@ -25,86 +24,11 @@ use serde::Deserialize;
 use serde::Serialize;
 
 use crate::error::ProxyResult;
-use crate::DEFAULT_PROXY_GRPC_PORT;
-use crate::DEFAULT_PROXY_REMOTING_PORT;
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum ProxyMode {
-    #[default]
-    Cluster,
-    Local,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(default, rename_all = "camelCase")]
-pub struct GrpcConfig {
-    pub listen_addr: String,
-    pub max_decoding_message_size: usize,
-    pub max_encoding_message_size: usize,
-    pub concurrency_limit_per_connection: usize,
-    pub use_endpoint_port_from_request: bool,
-}
-
-impl Default for GrpcConfig {
-    fn default() -> Self {
-        Self {
-            listen_addr: format!("0.0.0.0:{DEFAULT_PROXY_GRPC_PORT}"),
-            max_decoding_message_size: 8 * 1024 * 1024,
-            max_encoding_message_size: 8 * 1024 * 1024,
-            concurrency_limit_per_connection: 256,
-            use_endpoint_port_from_request: false,
-        }
-    }
-}
-
-impl GrpcConfig {
-    pub fn socket_addr(&self) -> ProxyResult<SocketAddr> {
-        self.listen_addr.parse().map_err(|error| {
-            RocketMQError::illegal_argument(format!(
-                "invalid proxy gRPC listen address '{}': {error}",
-                self.listen_addr
-            ))
-            .into()
-        })
-    }
-
-    pub fn listen_port(&self) -> ProxyResult<u16> {
-        Ok(self.socket_addr()?.port())
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(default, rename_all = "camelCase")]
-pub struct RemotingConfig {
-    pub enabled: bool,
-    pub listen_addr: String,
-}
-
-impl Default for RemotingConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            listen_addr: format!("0.0.0.0:{DEFAULT_PROXY_REMOTING_PORT}"),
-        }
-    }
-}
-
-impl RemotingConfig {
-    pub fn socket_addr(&self) -> ProxyResult<SocketAddr> {
-        self.listen_addr.parse().map_err(|error| {
-            RocketMQError::illegal_argument(format!(
-                "invalid proxy remoting listen address '{}': {error}",
-                self.listen_addr
-            ))
-            .into()
-        })
-    }
-
-    pub fn listen_port(&self) -> ProxyResult<u16> {
-        Ok(self.socket_addr()?.port())
-    }
-}
+pub use rocketmq_proxy_core::config::GrpcConfig;
+pub use rocketmq_proxy_core::config::ProxyMode;
+pub use rocketmq_proxy_core::config::RemotingConfig;
+pub use rocketmq_proxy_core::config::RuntimeConfig;
+pub use rocketmq_proxy_core::config::SessionConfig;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(default, rename_all = "camelCase")]
@@ -165,66 +89,6 @@ impl Default for LocalConfig {
             broker_listen_port: 10911,
             store_root_dir: "store/proxy/local-broker".to_owned(),
         }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(default, rename_all = "camelCase")]
-pub struct RuntimeConfig {
-    pub route_permits: usize,
-    pub producer_permits: usize,
-    pub consumer_permits: usize,
-    pub client_manager_permits: usize,
-}
-
-impl Default for RuntimeConfig {
-    fn default() -> Self {
-        Self {
-            route_permits: 512,
-            producer_permits: 1024,
-            consumer_permits: 1024,
-            client_manager_permits: 512,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(default, rename_all = "camelCase")]
-pub struct SessionConfig {
-    pub client_ttl_ms: u64,
-    pub receipt_handle_ttl_ms: u64,
-    pub auto_renew_enabled: bool,
-    pub min_long_polling_timeout_ms: u64,
-    pub max_long_polling_timeout_ms: u64,
-}
-
-impl Default for SessionConfig {
-    fn default() -> Self {
-        Self {
-            client_ttl_ms: 60_000,
-            receipt_handle_ttl_ms: 5 * 60_000,
-            auto_renew_enabled: true,
-            min_long_polling_timeout_ms: 5_000,
-            max_long_polling_timeout_ms: 20_000,
-        }
-    }
-}
-
-impl SessionConfig {
-    pub fn client_ttl(&self) -> Duration {
-        Duration::from_millis(self.client_ttl_ms.max(1))
-    }
-
-    pub fn receipt_handle_ttl(&self) -> Duration {
-        Duration::from_millis(self.receipt_handle_ttl_ms.max(1))
-    }
-
-    pub fn min_long_polling_timeout(&self) -> Duration {
-        Duration::from_millis(self.min_long_polling_timeout_ms)
-    }
-
-    pub fn max_long_polling_timeout(&self) -> Duration {
-        Duration::from_millis(self.max_long_polling_timeout_ms.max(self.min_long_polling_timeout_ms))
     }
 }
